+++ pkg/Microsoft.ML.AutoML/Microsoft.ML.AutoML.nupkgproj
@@ -10,8 +10,8 @@
     <ProjectReference Include="../Microsoft.ML.LightGbm/Microsoft.ML.LightGbm.nupkgproj" />
     <ProjectReference Include="../Microsoft.ML.Mkl.Components/Microsoft.ML.Mkl.Components.nupkgproj" />
     <ProjectReference Include="../Microsoft.ML.Recommender/Microsoft.ML.Recommender.nupkgproj" />
-    <ProjectReference Include="../Microsoft.ML/Microsoft.ML.Dnn.nupkgproj" />
-    <ProjectReference Include="../Microsoft.ML/Microsoft.ML.ImageAnalytics.nupkgproj" />
+    <ProjectReference Include="../Microsoft.ML.Vision/Microsoft.ML.Vision.nupkgproj" />
+    <ProjectReference Include="../Microsoft.ML.ImageAnalytics/Microsoft.ML.ImageAnalytics.nupkgproj" />
   </ItemGroup>
 
 </Project>
