--- pkg/Microsoft.ML.Dnn/Microsoft.ML.Dnn.nupkgproj
@@ -1,15 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk" DefaultTargets="Pack">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-    <PackageDescription>Microsoft.ML.Dnn contains APIs to do high level DNN training such as image classification.</PackageDescription>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="System.IO.FileSystem.AccessControl" Version="$(SystemIOFileSystemAccessControl)" />
-    <PackageReference Include="System.Security.Principal.Windows" Version="$(SystemSecurityPrincipalWindows)" />
-    <PackageReference Include="TensorFlow.NET" Version="$(TensorflowDotNETVersion)" />
-    <ProjectReference Include="../Microsoft.ML/Microsoft.ML.nupkgproj" />
-  </ItemGroup>
-
-</Project>
