--- pkg/Microsoft.ML.Dnn/Microsoft.ML.Dnn.symbols.nupkgproj
@@ -1,5 +1,0 @@
-<Project DefaultTargets="Pack">
-
-  <Import Project="Microsoft.ML.Dnn.nupkgproj" />
-
-</Project>
