+++ pkg/Microsoft.ML.Vision/Microsoft.ML.Vision.symbols.nupkgproj
@@ -1,0 +1,5 @@
+<Project DefaultTargets="Pack">
+
+  <Import Project="Microsoft.ML.Vision.nupkgproj" />
+
+</Project>
