+++ src/Microsoft.ML.AutoML/TrainerExtensions/MultiTrainerExtensions.cs
@@ -11,7 +11,7 @@ using Microsoft.ML.Trainers.LightGbm;
 
 namespace Microsoft.ML.AutoML
 {
-    using static Microsoft.ML.Dnn.ImageClassificationTrainer;
+    using static Microsoft.ML.Vision.ImageClassificationTrainer;
     using ITrainerEstimator = ITrainerEstimator<IPredictionTransformer<object>, object>;
 
     internal class AveragedPerceptronOvaExtension : ITrainerExtension
