+++ src/Microsoft.ML.CodeGenerator/Templates/Console/PredictProject.cs
@@ -43,8 +43,8 @@ namespace Microsoft.ML.CodeGenerator.Templates.Console
             this.Write("    <PackageReference Include=\"Microsoft.ML.ImageAnalytics\" Version=\"1.4.0\" />\r\n");
 }
  if (IncludeImageClassificationPackage){ 
-            this.Write("    <PackageReference Include=\"Microsoft.ML.Dnn\" Version=\"1.4.0\" />\r\n\t<PackageRef" +
-                    "erence Include=\"SciSharp.TensorFlow.Redist\" Version=\"1.14.0\" />\r\n");
+            this.Write("    <PackageReference Include=\"Microsoft.ML.Vision\" Version=\"1.4.0\" />\r\n\t<Package" +
+                    "Reference Include=\"SciSharp.TensorFlow.Redist\" Version=\"1.14.0\" />\r\n");
 }
             this.Write("  </ItemGroup>\r\n  <ItemGroup>\r\n    <ProjectReference Include=\"..\\");
             this.Write(this.ToStringHelper.ToStringWithCulture(Namespace));
