+++ src/Microsoft.ML.CodeGenerator/Templates/Console/PredictProject.tt
@@ -25,7 +25,7 @@
     <PackageReference Include="Microsoft.ML.ImageAnalytics" Version="1.4.0" />
 <#}#>
 <# if (IncludeImageClassificationPackage){ #>
-    <PackageReference Include="Microsoft.ML.Dnn" Version="1.4.0" />
+    <PackageReference Include="Microsoft.ML.Vision" Version="1.4.0" />
 	<PackageReference Include="SciSharp.TensorFlow.Redist" Version="1.14.0" />
 <#}#>
   </ItemGroup>
