+++ src/Microsoft.ML.StandardTrainers/Optimizer/LearningRateScheduler.cs
@@ -5,17 +5,18 @@
 using System;
 using System.Collections.Generic;
 
-namespace Microsoft.ML.Dnn
+namespace Microsoft.ML.Trainers
 {
     /// <summary>
     /// A class that contains the current train state to use for learning rate scheduling.
     /// </summary>
-    internal class TrainState
+    [BestFriend]
+    internal class DnnTrainState
     {
-        internal int CurrentBatchIndex;
-        internal int CurrentEpoch;
-        internal int BatchSize;
-        internal int BatchesPerEpoch;
+        public int CurrentBatchIndex;
+        public int CurrentEpoch;
+        public int BatchSize;
+        public int BatchesPerEpoch;
     }
 
     /// <summary>
@@ -23,11 +24,13 @@ namespace Microsoft.ML.Dnn
     /// </summary>
     public abstract class LearningRateScheduler
     {
+        [BestFriend]
         internal LearningRateScheduler()
         {
         }
 
-        internal abstract float GetLearningRate(TrainState options);
+        [BestFriend]
+        internal abstract float GetLearningRate(DnnTrainState options);
     }
 
     /// <summary>
@@ -75,7 +78,7 @@ namespace Microsoft.ML.Dnn
             List<LearningRateSchedulerItem> lrs = new List<LearningRateSchedulerItem>();
             int[] epochs = { 182, 136, 91, 0 };
             float[] scalingFactor = { 0.0001f, 0.01f, 0.1f, 1.0f };
-            for(int i = 0; i < 4; i++)
+            for (int i = 0; i < 4; i++)
             {
                 LearningRateSchedulerItem item = new LearningRateSchedulerItem(epochs[i], scalingFactor[i]);
                 lrs.Add(item);
@@ -106,7 +109,7 @@ namespace Microsoft.ML.Dnn
         /// </summary>
         private float GetLearningRateScheduleMultiplier(int epoch)
         {
-            for(int i = 0; i < _lrSchedule.Count; i++)
+            for (int i = 0; i < _lrSchedule.Count; i++)
             {
                 if (epoch >= _lrSchedule[i].Epoch)
                 {
@@ -119,7 +122,7 @@ namespace Microsoft.ML.Dnn
         /// <summary>
         /// This function returns the Learning rate using linear scale rule and LR decay.
         /// </summary>
-        internal override float GetLearningRate(TrainState trainstate)
+        internal override float GetLearningRate(DnnTrainState trainstate)
         {
             float learningrate;
             float initialLearningRate = BaseLearningRate * trainstate.BatchSize / 128;
@@ -183,14 +186,14 @@ namespace Microsoft.ML.Dnn
         /// <summary>
         /// Computes exponentially decayed learning rate
         /// </summary>
-        internal override float GetLearningRate(TrainState trainstate)
+        internal override float GetLearningRate(DnnTrainState trainstate)
         {
             int numSamplesPerEpoch = trainstate.BatchSize * trainstate.BatchesPerEpoch;
-            DecaySteps = (int) (numSamplesPerEpoch * NumEpochsPerDecay / trainstate.BatchSize);
-            GlobalStep = (trainstate.CurrentEpoch) *(trainstate.BatchesPerEpoch) + trainstate.CurrentBatchIndex;
+            DecaySteps = (int)(numSamplesPerEpoch * NumEpochsPerDecay / trainstate.BatchSize);
+            GlobalStep = (trainstate.CurrentEpoch) * (trainstate.BatchesPerEpoch) + trainstate.CurrentBatchIndex;
             float decayPower = (float)GlobalStep / DecaySteps;
-            decayPower = Staircase ? (float) Math.Floor(decayPower) : decayPower;
-            float decayedLearningRate = LearningRate * (float) Math.Pow(DecayRate, decayPower);
+            decayPower = Staircase ? (float)Math.Floor(decayPower) : decayPower;
+            float decayedLearningRate = LearningRate * (float)Math.Pow(DecayRate, decayPower);
             return decayedLearningRate;
         }
 
@@ -238,22 +241,22 @@ namespace Microsoft.ML.Dnn
             Cycle = cycle;
         }
 
-        internal override float GetLearningRate(TrainState trainstate)
+        internal override float GetLearningRate(DnnTrainState trainstate)
         {
             int numSamplesPerEpoch = trainstate.BatchSize * trainstate.BatchesPerEpoch;
-            int decaySteps = (int) (numSamplesPerEpoch * NumEpochsPerDecay / trainstate.BatchSize);
-            int globalStep = (trainstate.CurrentEpoch) *(trainstate.BatchesPerEpoch) + trainstate.CurrentBatchIndex;
+            int decaySteps = (int)(numSamplesPerEpoch * NumEpochsPerDecay / trainstate.BatchSize);
+            int globalStep = (trainstate.CurrentEpoch) * (trainstate.BatchesPerEpoch) + trainstate.CurrentBatchIndex;
 
             float decayedLearningRate;
             if (Cycle && globalStep > decaySteps)
             {
-                float calculatedStep = (float)decaySteps * (float)Math.Ceiling((double)globalStep/(double)decaySteps);
-                decayedLearningRate = (LearningRate - EndLearningRate) * ((float) Math.Pow((1 - (float)globalStep / calculatedStep), Power)) + EndLearningRate;
+                float calculatedStep = (float)decaySteps * (float)Math.Ceiling((double)globalStep / (double)decaySteps);
+                decayedLearningRate = (LearningRate - EndLearningRate) * ((float)Math.Pow((1 - (float)globalStep / calculatedStep), Power)) + EndLearningRate;
             }
             else
             {
                 float calculatedStep = Math.Min(globalStep, decaySteps);
-                decayedLearningRate = (LearningRate - EndLearningRate) * ((float) Math.Pow((1 - calculatedStep / (float)decaySteps), Power)) + EndLearningRate;
+                decayedLearningRate = (LearningRate - EndLearningRate) * ((float)Math.Pow((1 - calculatedStep / (float)decaySteps), Power)) + EndLearningRate;
             }
             return decayedLearningRate;
         }
