+++ src/Microsoft.ML.TensorFlow/TensorFlowModel.cs
@@ -3,7 +3,7 @@
 // See the LICENSE file in the project root for more information.
 
 using Microsoft.ML.Runtime;
-using Microsoft.ML.Transforms.TensorFlow;
+using Microsoft.ML.TensorFlow;
 using Tensorflow;
 
 namespace Microsoft.ML.Transforms
