+++ src/Microsoft.ML.TensorFlow/TensorTypeExtensions.cs
@@ -4,12 +4,9 @@
 
 using System;
 using Microsoft.ML.Internal.Utilities;
-using NumSharp.Backends;
-using NumSharp.Backends.Unmanaged;
-using NumSharp.Utilities;
 using Tensorflow;
 
-namespace Microsoft.ML.Transforms
+namespace Microsoft.ML.TensorFlow
 {
     [BestFriend]
     internal static class TensorTypeExtensions
